requirements.txt
python-bsonjs==0.3.0
Flask==2.2.3
Flask-PyMongo==2.3.0
pymongo==4.6.1